requirements.txt
pwntools
pyradamsa
python-nmap
scapy
